Stopped enemy from losing limbs from non weapon contact
