+++ Assets/Scenes/Main Map.unity
@@ -17381,7 +17381,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   player: {fileID: 1826147901}
-  AttackDistance: 2.5
+  AttackDistance: 1
 --- !u!1001 &2012760234
 PrefabInstance:
   m_ObjectHideFlags: 0
